Remove any intentional panics in functions creating a database.
In a production use, there shouldnt be panics in libraries that would possibly result in the whole application closing. Although there may be panics as a result of an internal error, this specific library should leave it up to the developer on handling how to end their application as a result of an error/panic of specific functions in unqlite.rs.